test(platform-wallet): assert the poll-loop failure report fires once per wait
The existing test pinned the report flag's final state, not the suppression
the flag exists to provide. Deleting the `if !*reported` guard so every poll
iteration logs still left `reported == true`, so the test passed and the
regression would have shipped silently. Log volume from inside a poll loop is
the whole point of the guard, and nothing was measuring it.

The new test drives three iterations of one wait against a permanently
failing persister and asserts EXACTLY ONE error event, counting matching
events rather than observing that one exists. An assertion that merely finds
a report present is satisfied just as happily by one per iteration.

Mutation check, as required:
  guard deleted  -> poll_read_reports_a_permanent_failure_once_per_wait_not_once_per_iteration
                    FAILS: "three iterations of one wait must produce exactly
                    one report, got 3" (left: 3, right: 1)
                 -> 952 passed, 1 failed: the new test is the ONLY one that
                    changes colour, so it isolates the suppression property.
                    Notably poll_read_degrades_to_a_miss_on_permanent_backend_errors
                    stays green under the mutation, which is the direct
                    evidence that the flag-state assertion never covered this.
  guard restored -> green.

Capturing the events needed the recorder harness that already existed in
`wallet_lifecycle`'s test module, so it moves to `test_support` and both call
sites share it. Moved verbatim: a second harness would have to re-derive the
same constraint, and the naive alternative is a trap — a per-test
`set_default` swap races tracing's process-global callsite interest cache
under the parallel harness. The harness stays `#[cfg(test)]` because
`tracing-subscriber` is a dev-dependency. `wallet_lifecycle`'s own test is
unchanged and keeps its full strength (warn present AND error absent).

No production code changed.

Verified (`--no-deps` required: pre-existing unrelated rs-drive import):
  clippy --no-deps -p platform-wallet -p platform-wallet-ffi --all-targets
    -D warnings                                          exit 0
  test -p platform-wallet -p platform-wallet-ffi, CLAUDIUS_FORCE=1, twice
                                                         exit 0, 1324 each
