+++ packages/rs-platform-wallet/src/manager/wallet_lifecycle.rs
@@ -1295,18 +1295,14 @@ mod persist_retry_tests {
     //! Registration-path persistence: single-attempt `store` with typed
     //! error propagation, bounded `load` retry, and log-level policy.
 
-    use std::cell::RefCell;
     use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::{Arc, Mutex, OnceLock};
+    use std::sync::Arc;
     use std::time::Duration;
 
     use key_wallet::mnemonic::{Language, Mnemonic};
     use key_wallet::wallet::initialization::WalletAccountCreationOptions;
     use key_wallet::Network;
-    use tracing::field::{Field, Visit};
     use tracing::Level;
-    use tracing_subscriber::layer::{Context, SubscriberExt};
-    use tracing_subscriber::Layer;
 
     use crate::changeset::{
         ClientStartState, PersistenceError, PersistenceErrorKind, PlatformWalletChangeSet,
@@ -1333,93 +1329,7 @@ mod persist_retry_tests {
         PersistenceError::backend_with_kind(PersistenceErrorKind::Fatal, "simulated corruption")
     }
 
-    /// Captures the level and message of every `tracing` event recorded
-    /// while registered as the active recorder for the current thread (see
-    /// [`RecordingGuard`]).
-    #[derive(Clone, Default)]
-    struct RecordedEvents(Arc<Mutex<Vec<(Level, String)>>>);
-
-    impl RecordedEvents {
-        fn entries(&self) -> Vec<(Level, String)> {
-            self.0.lock().expect("recorded events mutex").clone()
-        }
-
-        fn record(&self, event: &tracing::Event<'_>) {
-            struct MessageVisitor(String);
-            impl Visit for MessageVisitor {
-                fn record_debug(&mut self, field: &Field, value: &dyn std::fmt::Debug) {
-                    if field.name() == "message" {
-                        self.0 = format!("{value:?}");
-                    }
-                }
-            }
-            let mut visitor = MessageVisitor(String::new());
-            event.record(&mut visitor);
-            self.0
-                .lock()
-                .expect("recorded events mutex")
-                .push((*event.metadata().level(), visitor.0));
-        }
-    }
-
-    thread_local! {
-        /// The [`RecordedEvents`] a test on THIS thread wants routed to it,
-        /// if any. Set/cleared only by [`RecordingGuard`].
-        static ACTIVE_RECORDER: RefCell<Option<RecordedEvents>> = const { RefCell::new(None) };
-    }
-
-    /// Routes every event to whichever [`RecordedEvents`] is registered for
-    /// the emitting thread, via [`ACTIVE_RECORDER`]. Installed as the
-    /// process-wide default exactly once — never per-test.
-    ///
-    /// A per-test `tracing::subscriber::set_default` swap is flaky under
-    /// `cargo test`'s parallel harness: tracing's per-callsite `Interest`
-    /// cache is process-global, and a concurrently-running test's own
-    /// swap/drop can race the interest rebuild your swap triggers, so the
-    /// event silently never reaches your subscriber even though dispatch
-    /// itself stays correctly on your own thread (confirmed: the emitting
-    /// thread ID matched the installing thread ID on a captured failure).
-    /// Installing the routing subscriber once, before any callsite is ever
-    /// hit, sidesteps the race — routing then happens through an ordinary
-    /// thread-local this code owns, not through tracing's default-swap
-    /// machinery.
-    struct RecorderRouter;
-
-    impl<S: tracing::Subscriber> Layer<S> for RecorderRouter {
-        fn on_event(&self, event: &tracing::Event<'_>, _ctx: Context<'_, S>) {
-            ACTIVE_RECORDER.with(|slot| {
-                if let Some(recorder) = slot.borrow().as_ref() {
-                    recorder.record(event);
-                }
-            });
-        }
-    }
-
-    static GLOBAL_ROUTER_INIT: OnceLock<()> = OnceLock::new();
-
-    /// Scopes [`ACTIVE_RECORDER`] to `recorder` for the current thread, for
-    /// the guard's lifetime.
-    struct RecordingGuard;
-
-    impl RecordingGuard {
-        fn install(recorder: RecordedEvents) -> Self {
-            GLOBAL_ROUTER_INIT.get_or_init(|| {
-                let subscriber = tracing_subscriber::registry().with(RecorderRouter);
-                // Another thread may have already won this race; either
-                // way, the routing subscriber is the process-wide default
-                // by the time `get_or_init` returns to any caller.
-                let _ = tracing::subscriber::set_global_default(subscriber);
-            });
-            ACTIVE_RECORDER.with(|slot| *slot.borrow_mut() = Some(recorder));
-            Self
-        }
-    }
-
-    impl Drop for RecordingGuard {
-        fn drop(&mut self) {
-            ACTIVE_RECORDER.with(|slot| *slot.borrow_mut() = None);
-        }
-    }
+    use crate::test_support::tracing_capture::{RecordedEvents, RecordingGuard};
 
     /// Persister whose `store` / `flush` / `load` outcomes are scripted so
     /// the registration path can be driven deterministically.
